Split up components in InvestRedeem
